requirements.txt
asgiref==3.3.1
certifi==2020.11.8
cffi==1.14.3
chardet==3.0.4
cryptography==3.2.1
defusedxml==0.6.0
Django==3.1.3
django-allauth==0.43.0
django-crispy-forms==1.9.2
django-filter==2.4.0
django-polymorphic==3.0.0
docxtpl==0.11.3
idna==2.10
Jinja2==2.11.3
lxml==4.6.2
MarkupSafe==1.1.1
oauthlib==3.1.0
Pillow==8.0.1
pycparser==2.20
PyJWT==1.7.1
python-docx==0.8.10
python-dotenv==0.15.0
python3-openid==3.2.0
pytz==2020.4
requests==2.24.0
requests-oauthlib==1.3.0
six==1.15.0
sorl-thumbnail==12.7.0
sqlparse==0.4.1
stripe==2.55.0
urllib3==1.25.11
